refactor: split app/setup.rs by concern
setup.rs had grown to 1828 lines, the largest file in the repository, and
held several concerns that had nothing to do with one another.

R source resolution becomes app/setup/, divided into the resolution core,
the r_source_overrides providers, and R_HOME discovery with rig
integration, with script execution alongside them. Sourcing .Rprofile and
building a reedline history session id move out to app/ entirely, since
neither has anything to do with setting up R; their callers now address
them there rather than through setup, so the boundary is real rather than
a re-export.

Each test module moves next to the code it covers. No behaviour changes,
and the test count is unchanged at 1149.
