+++ crates/arf-console/src/app/setup/overrides.rs
@@ -1,479 +1,14 @@
-//! R setup, script execution mode, and session ID creation.
+//! Directory-level R source overrides.
 
-use crate::app::config_load::load_config_or_warn;
-use crate::cli::Cli;
-use crate::config;
-use crate::config::{
-    Config, RSource, RSourceMode, RSourceOverride, RSourceOverrideInfo, RSourceStatus,
-};
+use super::r_source::RSourceResolutionReport;
+use super::rig::{ResolvedRSource, setup_r_via_selected_rig_version};
+use crate::config::{RSourceOverride, RSourceOverrideInfo, RSourceStatus};
 use crate::external;
 use crate::rversion;
-use anyhow::{Context, Result};
-use reedline::Reedline;
-use std::fs;
+use anyhow::Result;
 use std::path::{Path, PathBuf};
 
-/// Run in script execution mode (non-interactive).
-pub(crate) fn run_script(cli: &Cli) -> Result<()> {
-    // Load configuration (from file or default)
-    let config = load_config_or_warn(cli.config.as_ref());
-
-    // Set up R based on r_source config (with optional CLI override)
-    let resolution = setup_r(
-        &config.startup.r_source,
-        &config.experimental.r_source_overrides,
-        None,
-        cli.r_home.as_deref(),
-        cli.r_version.as_deref(),
-        cli.no_r_source_overrides,
-    )?;
-    resolution.emit_diagnostics();
-    if let Some(notice) = script_override_notice(&resolution) {
-        eprintln!("{notice}");
-    }
-
-    // Ensure LD_LIBRARY_PATH includes R library directory
-    if let Err(e) = arf_libr::ensure_ld_library_path() {
-        log::warn!("Could not set LD_LIBRARY_PATH: {}", e);
-    }
-
-    // Generate R initialization arguments from CLI flags
-    let r_args = cli.r_args();
-    let r_args_refs: Vec<&str> = r_args.iter().map(|s| s.as_str()).collect();
-
-    // Initialize R with CLI-specified flags
-    unsafe {
-        arf_libr::initialize_r_with_args(&r_args_refs).context("Failed to initialize R")?;
-    }
-
-    // Source R profile files (Windows only)
-    #[cfg(windows)]
-    source_r_profiles(&r_args);
-
-    // Get the code to execute
-    let code = if let Some(eval_code) = &cli.eval {
-        eval_code.clone()
-    } else if let Some(script_path) = cli.script_file() {
-        if script_path == std::path::Path::new("-") {
-            use std::io::Read;
-            let mut buf = String::new();
-            std::io::stdin()
-                .read_to_string(&mut buf)
-                .context("Failed to read from stdin")?;
-            buf
-        } else {
-            fs::read_to_string(script_path)
-                .with_context(|| format!("Failed to read script file: {}", script_path.display()))?
-        }
-    } else {
-        // Should not happen - we checked script_mode earlier
-        return Ok(());
-    };
-
-    // Evaluate the code - use reprex mode if enabled (CLI or config)
-    let reprex_enabled = cli.reprex || config.startup.mode.reprex;
-    if reprex_enabled {
-        // In reprex mode, echo source code before each result
-        match arf_harp::eval_string_reprex(&code, &config.mode.reprex.comment) {
-            Ok(_) => Ok(()),
-            Err(e) => {
-                eprintln!("{}", e);
-                Ok(())
-            }
-        }
-    } else {
-        // Normal script execution
-        match arf_harp::eval_string(&code) {
-            Ok(_) => Ok(()),
-            Err(e) => {
-                eprintln!("{}", e);
-                Ok(())
-            }
-        }
-    }
-}
-
-/// Set up R based on r_source configuration.
-///
-/// CLI options override config in this order:
-/// 1. `cli_r_home` - explicit R_HOME path
-/// 2. `cli_version` - rig version specification
-/// 3. Experimental directory-level R source overrides
-/// 4. Config `r_source` setting
-///
-/// The state of override resolution for reporting to users and headless clients.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum RSourceOverrideState {
-    /// An override selected the active R installation.
-    Applied,
-    /// No R source override providers are configured.
-    NotConfigured,
-    /// No configured provider matched a file in the current directory.
-    NoMatch,
-    /// At least one provider was evaluated but no override could be applied.
-    Failed,
-    /// Override resolution was disabled by the CLI.
-    Disabled,
-    /// A CLI R source took precedence over overrides.
-    ShadowedByCli,
-}
-
-impl RSourceOverrideState {
-    pub(crate) fn as_str(self) -> &'static str {
-        match self {
-            Self::Applied => "applied",
-            Self::NotConfigured => "not_configured",
-            Self::NoMatch => "no_match",
-            Self::Failed => "failed",
-            Self::Disabled => "disabled",
-            Self::ShadowedByCli => "shadowed_by_cli",
-        }
-    }
-}
-
-/// The result of resolving the configured R source and any directory override.
-#[derive(Debug, Clone)]
-pub(crate) struct RSourceResolutionReport {
-    pub(crate) status: RSourceStatus,
-    pub(crate) r_home: Option<PathBuf>,
-    pub(crate) provider: Option<String>,
-    pub(crate) file: Option<std::path::PathBuf>,
-    pub(crate) key: Option<String>,
-    pub(crate) requested_version: Option<String>,
-    pub(crate) resolved_version: Option<String>,
-    pub(crate) diagnostics: Vec<String>,
-    pub(crate) override_state: RSourceOverrideState,
-}
-
-impl RSourceResolutionReport {
-    fn from_status(
-        status: RSourceStatus,
-        r_home: Option<PathBuf>,
-        override_state: RSourceOverrideState,
-    ) -> Self {
-        Self {
-            status,
-            r_home,
-            provider: None,
-            file: None,
-            key: None,
-            requested_version: None,
-            resolved_version: None,
-            diagnostics: Vec::new(),
-            override_state,
-        }
-    }
-
-    fn applied(
-        status: RSourceStatus,
-        r_home: PathBuf,
-        info: RSourceOverrideInfo,
-        diagnostics: Vec<String>,
-    ) -> Self {
-        Self {
-            status,
-            r_home: Some(r_home),
-            provider: Some(info.provider),
-            file: info.file,
-            key: info.key,
-            requested_version: Some(info.requested_version),
-            resolved_version: Some(info.resolved_version),
-            diagnostics,
-            override_state: RSourceOverrideState::Applied,
-        }
-    }
-
-    /// Emit all override diagnostics as one warning block for normal startup.
-    pub(crate) fn emit_diagnostics(&self) {
-        if self.diagnostics.is_empty() {
-            return;
-        }
-        let block = self.diagnostics.join("\n");
-        eprintln!("{block}");
-        log::warn!("{block}");
-    }
-
-    pub(crate) fn override_info(&self) -> Option<RSourceOverrideInfo> {
-        Some(RSourceOverrideInfo {
-            provider: self.provider.clone()?,
-            file: self.file.clone(),
-            key: self.key.clone(),
-            requested_version: self.requested_version.clone()?,
-            resolved_version: self.resolved_version.clone()?,
-        })
-    }
-}
-
-/// Resolve R and return a report without changing the process environment.
-pub(crate) fn resolve_r_source(
-    r_source: &RSource,
-    r_source_overrides: &[RSourceOverride],
-    base_dir: Option<&Path>,
-    cli_r_home: Option<&std::path::Path>,
-    cli_version: Option<&str>,
-    no_r_source_overrides: bool,
-) -> Result<RSourceResolutionReport> {
-    // CLI --r-home overrides everything. An explicit disable still controls
-    // the reported override state without changing the selected R source.
-    if let Some(path) = cli_r_home {
-        if !path.exists() {
-            anyhow::bail!(
-                "R_HOME path does not exist: {}\n\
-                 Check your --r-home argument.",
-                path.display()
-            );
-        }
-        // Resolve R_HOME: if path looks like an installation prefix (has bin/R),
-        // run `bin/R RHOME` to get the actual R_HOME directory
-        let r_home = resolve_r_home_from_path(path)?;
-        log::info!("Using R from --r-home: {}", r_home.display());
-        return Ok(RSourceResolutionReport::from_status(
-            RSourceStatus::ExplicitPath {
-                path: r_home.clone(),
-            },
-            Some(r_home),
-            if no_r_source_overrides {
-                RSourceOverrideState::Disabled
-            } else {
-                RSourceOverrideState::ShadowedByCli
-            },
-        ));
-    }
-
-    // CLI --with-r-version overrides config (uses rig)
-    if let Some(version) = cli_version {
-        return setup_r_via_rig(version).map(|resolved| {
-            RSourceResolutionReport::from_status(
-                resolved.status,
-                resolved.r_home,
-                if no_r_source_overrides {
-                    RSourceOverrideState::Disabled
-                } else {
-                    RSourceOverrideState::ShadowedByCli
-                },
-            )
-        });
-    }
-
-    if no_r_source_overrides {
-        return setup_r_fallback(r_source, RSourceOverrideState::Disabled);
-    }
-
-    if let Some(result) = setup_r_via_overrides(r_source_overrides, base_dir) {
-        match result {
-            OverrideResolution::Applied {
-                status,
-                r_home,
-                info,
-                diagnostics,
-            } => {
-                return Ok(RSourceResolutionReport::applied(
-                    *status,
-                    r_home,
-                    info,
-                    diagnostics,
-                ));
-            }
-            OverrideResolution::Fallback { diagnostics } => {
-                let mut report = setup_r_fallback(r_source, RSourceOverrideState::Failed)?;
-                report.diagnostics = diagnostics;
-                return Ok(report);
-            }
-        }
-    }
-
-    let override_state = if r_source_overrides.is_empty() {
-        RSourceOverrideState::NotConfigured
-    } else {
-        RSourceOverrideState::NoMatch
-    };
-    setup_r_fallback(r_source, override_state)
-}
-
-/// Apply a resolved R source to the process environment during startup.
-pub(crate) fn apply_r_source(resolution: &RSourceResolutionReport) -> Result<()> {
-    if matches!(resolution.status, RSourceStatus::Path) {
-        return Ok(());
-    }
-
-    let r_home = resolution
-        .r_home
-        .as_ref()
-        .context("Resolved R source has no R_HOME")?;
-    // SAFETY: This is called single-threaded during startup, before R is initialized.
-    unsafe { std::env::set_var("R_HOME", r_home) };
-    Ok(())
-}
-
-/// Set up R and return a resolution report for display and feature gating.
-pub(crate) fn setup_r(
-    r_source: &RSource,
-    r_source_overrides: &[RSourceOverride],
-    base_dir: Option<&Path>,
-    cli_r_home: Option<&std::path::Path>,
-    cli_version: Option<&str>,
-    no_r_source_overrides: bool,
-) -> Result<RSourceResolutionReport> {
-    let resolution = resolve_r_source(
-        r_source,
-        r_source_overrides,
-        base_dir,
-        cli_r_home,
-        cli_version,
-        no_r_source_overrides,
-    )?;
-    apply_r_source(&resolution)?;
-    Ok(resolution)
-}
-
-fn setup_r_fallback(
-    r_source: &RSource,
-    override_state: RSourceOverrideState,
-) -> Result<RSourceResolutionReport> {
-    setup_r_fallback_with(
-        r_source,
-        override_state,
-        external::rig::rig_available,
-        external::rig::resolve_version,
-    )
-}
-
-fn setup_r_fallback_with<FAvailable, FResolve>(
-    r_source: &RSource,
-    override_state: RSourceOverrideState,
-    rig_available: FAvailable,
-    resolve_version: FResolve,
-) -> Result<RSourceResolutionReport>
-where
-    FAvailable: Fn() -> bool,
-    FResolve:
-        Fn(&str) -> std::result::Result<external::rig::ResolvedVersion, external::rig::RigError>,
-{
-    let resolved = match r_source {
-        RSource::Mode(RSourceMode::Auto) => {
-            // Auto mode: try rig if available, otherwise use PATH
-            if rig_available() {
-                match resolve_version("default") {
-                    Ok(resolved) => {
-                        log::info!("Using rig default R version: {}", resolved.version);
-                        ResolvedRSource {
-                            status: RSourceStatus::Rig {
-                                version: resolved.version,
-                                override_info: None,
-                            },
-                            r_home: Some(PathBuf::from(resolved.r_home)),
-                        }
-                    }
-                    Err(e) => {
-                        log::debug!("Could not get rig default version: {}", e);
-                        log::info!("Using R from PATH");
-                        ResolvedRSource {
-                            // Do not discover or assign R_HOME here. R will discover it
-                            // during initialization, preserving PATH-mode startup behavior.
-                            status: RSourceStatus::Path,
-                            r_home: None,
-                        }
-                    }
-                }
-            } else {
-                log::info!("Using R from PATH (rig not available)");
-                ResolvedRSource {
-                    // Do not discover or assign R_HOME here. R will discover it
-                    // during initialization, preserving PATH-mode startup behavior.
-                    status: RSourceStatus::Path,
-                    r_home: None,
-                }
-            }
-        }
-        RSource::Mode(RSourceMode::Rig) => {
-            // Rig mode: require rig
-            if !rig_available() {
-                anyhow::bail!(
-                    r#"r_source = "rig" but rig is not installed.
-Install rig from https://github.com/r-lib/rig or use "auto"."#
-                );
-            }
-            match resolve_version("default") {
-                Ok(resolved) => {
-                    log::info!("Using rig default R version: {}", resolved.version);
-                    ResolvedRSource {
-                        status: RSourceStatus::Rig {
-                            version: resolved.version,
-                            override_info: None,
-                        },
-                        r_home: Some(PathBuf::from(resolved.r_home)),
-                    }
-                }
-                Err(e) => {
-                    anyhow::bail!("Failed to get rig default R version: {}", e);
-                }
-            }
-        }
-        RSource::Path { path } => {
-            // Explicit path mode
-            if !path.exists() {
-                anyhow::bail!(
-                    "R_HOME path does not exist: {}\n\
-                     Check your r_source configuration.",
-                    path.display()
-                );
-            }
-            log::info!("Using R from explicit path: {}", path.display());
-            ResolvedRSource {
-                status: RSourceStatus::ExplicitPath { path: path.clone() },
-                r_home: Some(path.clone()),
-            }
-        }
-    };
-
-    Ok(RSourceResolutionReport::from_status(
-        resolved.status,
-        resolved.r_home,
-        override_state,
-    ))
-}
-
-#[derive(Debug)]
-struct ResolvedRSource {
-    status: RSourceStatus,
-    r_home: Option<PathBuf>,
-}
-
-fn resolve_path_r_home_with<F>(find_r_library: F) -> Result<PathBuf>
-where
-    F: FnOnce() -> arf_libr::RResult<PathBuf>,
-{
-    let library_path = find_r_library().context("Failed to discover the R library from PATH")?;
-    arf_libr::r_home_from_library_path(&library_path).with_context(|| {
-        format!(
-            "Failed to derive R_HOME from discovered R library: {}",
-            library_path.display()
-        )
-    })
-}
-
-/// Resolve the PATH-mode R_HOME only for callers that need to report it.
-pub(crate) fn resolve_path_r_home_for_report(report: &mut RSourceResolutionReport) {
-    resolve_path_r_home_for_report_with(report, arf_libr::find_r_library);
-}
-
-fn resolve_path_r_home_for_report_with<F>(report: &mut RSourceResolutionReport, find_r_library: F)
-where
-    F: FnOnce() -> arf_libr::RResult<PathBuf>,
-{
-    if !matches!(report.status, RSourceStatus::Path) || report.r_home.is_some() {
-        return;
-    }
-
-    match resolve_path_r_home_with(find_r_library) {
-        Ok(r_home) => report.r_home = Some(r_home),
-        Err(error) => report.diagnostics.push(format!(
-            "Warning: Failed to determine R_HOME from PATH: {error}"
-        )),
-    }
-}
-
-enum OverrideResolution {
+pub(super) enum OverrideResolution {
     Applied {
         status: Box<RSourceStatus>,
         r_home: PathBuf,
@@ -486,7 +21,7 @@ enum OverrideResolution {
 }
 
 /// Try the configured directory-level R source overrides in priority order.
-fn setup_r_via_overrides(
+pub(super) fn setup_r_via_overrides(
     overrides: &[RSourceOverride],
     base_dir: Option<&Path>,
 ) -> Option<OverrideResolution> {
@@ -777,7 +312,7 @@ fn invalid_override_file_warning(provider: &str, file: &Path) -> String {
     )
 }
 
-fn script_override_notice(resolution: &RSourceResolutionReport) -> Option<String> {
+pub(super) fn script_override_notice(resolution: &RSourceResolutionReport) -> Option<String> {
     resolution
         .override_info()
         .map(|info| format!("# R source override: {}", info.display()))
@@ -841,209 +376,15 @@ fn not_installed_warning(
     }
 }
 
-/// Resolve R_HOME from a user-provided path.
-///
-/// The path can be either:
-/// - An installation prefix (e.g., `/opt/R/4.5.2`) containing `bin/R`
-/// - The actual R_HOME directory (e.g., `/opt/R/4.5.2/lib/R`)
-///
-/// If the path contains `bin/R`, we run it with `RHOME` to get the actual R_HOME.
-fn resolve_r_home_from_path(path: &std::path::Path) -> Result<std::path::PathBuf> {
-    // Check if this looks like an installation prefix (has bin/R)
-    let r_binary = path.join("bin").join("R");
-    if r_binary.exists() {
-        // Run `bin/R RHOME` to get the actual R_HOME
-        let output = std::process::Command::new(&r_binary)
-            .arg("RHOME")
-            .output()
-            .with_context(|| format!("Failed to run {} RHOME", r_binary.display()))?;
-
-        if !output.status.success() {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            anyhow::bail!("{} RHOME failed: {}", r_binary.display(), stderr);
-        }
-
-        let r_home = String::from_utf8_lossy(&output.stdout).trim().to_string();
-        if r_home.is_empty() {
-            anyhow::bail!("{} RHOME returned empty result", r_binary.display());
-        }
-
-        log::debug!(
-            "Resolved R_HOME from installation prefix: {} -> {}",
-            path.display(),
-            r_home
-        );
-        return Ok(std::path::PathBuf::from(r_home));
-    }
-
-    // Assume the path is already R_HOME
-    // Validate by checking for etc/Renviron
-    let renviron = path.join("etc").join("Renviron");
-    if !renviron.exists() {
-        log::warn!(
-            "Path {} does not look like R_HOME (missing etc/Renviron). \
-             Consider providing the installation prefix instead.",
-            path.display()
-        );
-    }
-
-    Ok(path.to_path_buf())
-}
-
-/// Resolve R via rig with a specific version (used for CLI --with-r-version).
-fn setup_r_via_rig(version_spec: &str) -> Result<ResolvedRSource> {
-    if !external::rig::rig_available() {
-        anyhow::bail!(
-            "--with-r-version requires rig to be installed.\n\
-             Install rig from https://github.com/r-lib/rig"
-        );
-    }
-
-    match external::rig::resolve_version(version_spec) {
-        Ok(resolved) => resolve_rig_resolution(resolved),
-        Err(e) => {
-            anyhow::bail!("Failed to resolve R version '{}': {}", version_spec, e);
-        }
-    }
-}
-
-/// Set up R from the exact semantic version selected for an override.
-///
-/// The override resolver has already selected a version from rig's reported
-/// version fields. Re-resolving its string would allow a rig name or alias to
-/// select a different installation.
-fn setup_r_via_selected_rig_version(
-    selected: &semver::Version,
-    versions: &[external::rig::RigVersion],
-) -> Result<ResolvedRSource> {
-    match external::rig::resolve_selected_version_from_versions(selected, versions) {
-        Ok(resolved) => resolve_rig_resolution(resolved),
-        Err(error) => anyhow::bail!("Failed to resolve R version '{}': {}", selected, error),
-    }
-}
-
-fn resolve_rig_resolution(resolved: external::rig::ResolvedVersion) -> Result<ResolvedRSource> {
-    log::info!(
-        "Using R version {} from {}",
-        resolved.version,
-        resolved.r_home
-    );
-    Ok(ResolvedRSource {
-        status: RSourceStatus::Rig {
-            version: resolved.version,
-            override_info: None,
-        },
-        r_home: Some(PathBuf::from(resolved.r_home)),
-    })
-}
-
-/// Source R profile files after R initialization.
-///
-/// This handles loading of:
-/// - Site-level Rprofile.site (unless --no-site-file or --vanilla)
-/// - User-level .Rprofile (unless --no-init-file or --vanilla)
-///
-/// On Windows, R's built-in profile loading is disabled during initialization
-/// for compatibility with `globalCallingHandlers()`, so we must manually
-/// source these files here.
-#[cfg(windows)]
-pub(crate) fn source_r_profiles(r_args: &[String]) {
-    // Fix .Platform$GUI before any R profiles or packages are loaded.
-    // See: https://github.com/eitsupi/arf/issues/168
-    arf_harp::override_platform_gui();
-
-    // Get R_HOME from environment (set earlier in setup_r)
-    let r_home = match std::env::var("R_HOME") {
-        Ok(path) => PathBuf::from(path),
-        Err(_) => {
-            log::warn!("R_HOME not set, skipping R profile sourcing");
-            return;
-        }
-    };
-
-    // Source site-level R profile unless --no-site-file or --vanilla
-    if !arf_harp::should_ignore_site_r_profile(r_args) {
-        arf_harp::source_site_r_profile(&r_home);
-    } else {
-        log::trace!("Skipping site R profile (--no-site-file or --vanilla)");
-    }
-
-    // Source user-level R profile unless --no-init-file or --vanilla
-    if !arf_harp::should_ignore_user_r_profile(r_args) {
-        arf_harp::source_user_r_profile();
-    } else {
-        log::trace!("Skipping user R profile (--no-init-file or --vanilla)");
-    }
-
-    // Call .First() then .First.sys() to match R's documented startup sequence
-    // (see `?Startup`). After profiles are loaded:
-    //   1. .First()     — user hook defined in .Rprofile (e.g. vscode-R session watcher)
-    //   2. .First.sys() — base package hook that loads default packages (utils, grDevices, ...)
-    // On Windows we source profiles manually (profiles disabled in setup_Rmainloop for
-    // globalCallingHandlers compatibility), so we must call these hooks manually too.
-    arf_harp::call_dot_first();
-    arf_harp::call_dot_first_sys();
-}
-
-/// Generate a history session ID when history is enabled and a history directory
-/// is available, or `None` otherwise.
-///
-/// This ensures IPC/session JSON does not misleadingly advertise history isolation
-/// when no history backend is configured.
-pub(crate) fn create_session_id(config: &Config) -> Option<reedline::HistorySessionId> {
-    if config.history.disabled {
-        return None;
-    }
-    // Check that a history directory is actually resolvable, matching the logic
-    // in Repl::r_history_path() / shell_history_path().
-    if config.history.dir.is_none() && config::history_dir().is_none() {
-        return None;
-    }
-    Reedline::create_history_session_id()
-}
-
-#[cfg(test)]
-mod session_id_tests {
-    use super::*;
-
-    #[test]
-    fn test_create_session_id_when_history_enabled() {
-        let mut config = Config::default();
-        // Ensure a history dir is available by setting it explicitly
-        config.history.dir = Some(std::env::temp_dir());
-        assert!(!config.history.disabled);
-        let id = create_session_id(&config);
-        assert!(
-            id.is_some(),
-            "should generate session ID when history is enabled"
-        );
-    }
-
-    #[test]
-    fn test_create_session_id_when_history_disabled() {
-        let mut config = Config::default();
-        config.history.disabled = true;
-        let id = create_session_id(&config);
-        assert!(id.is_none(), "should be None when history is disabled");
-    }
-
-    #[test]
-    fn test_create_session_id_respects_default_history_dir() {
-        // With default config (history.dir = None), session ID depends on
-        // whether the platform provides a data directory via history_dir().
-        let config = Config::default();
-        assert!(!config.history.disabled);
-        assert!(config.history.dir.is_none());
-        let id = create_session_id(&config);
-        // On most platforms history_dir() returns Some, so session ID is generated.
-        // On exotic platforms where it returns None, session ID should be None.
-        assert_eq!(id.is_some(), config::history_dir().is_some());
-    }
-}
-
 #[cfg(test)]
 mod r_source_override_tests {
+    use super::super::r_source::{
+        RSourceOverrideState, RSourceResolutionReport, apply_r_source,
+        resolve_path_r_home_for_report_with, resolve_r_source, setup_r_fallback_with,
+    };
+    use super::super::rig::ResolvedRSource;
     use super::*;
+    use crate::config::{Config, RSource, RSourceMode};
     use std::path::Path;
     use std::path::PathBuf;
     use std::sync::Arc;
